refactor(be): split shared inbound-DKIM types into types::email_challenge
The seven types that describe the inbound-DKIM challenge primitive
(EmailChallenge, EmailChallengeDnsInput, EmailChallengeSubmitDkimLeafArg,
EmailChallengeResolveViaDohArg, EmailChallengeError, EmailChallengeStatus,
EmailChallengeDiagnostics) plus their supporting enums (VerificationPath,
DohFailureReason) lived in types::email_recovery — same shape of
naming asymmetry we fixed on the canister side. Move them to a new
types::email_challenge module so the import path matches the type
name.

types::email_recovery is now slim: just EmailRecoveryCredential and
EmailRecoveryGetDelegationArgs — both genuinely recovery-specific.

Pure rename + module split, no candid surface change.
